added part 5 and modified part 4 blogTest

+++ part4/blogList/backend/tests/blog_api.test.js
@@ -9,17 +9,23 @@ import {initialBlogs, blogsInDb, usersInDb} from "./test_helper.js";
 const api = supertest(app);
 const baseUrl = "/api/blogs";
 
-beforeEach(async () => {
-	await Blog.deleteMany({});
-	const users = await usersInDb()
-		
-	for (let blog of initialBlogs) {
-		let blogObject = new Blog({...blog, user: users[0].id});
-		await blogObject.save();		
-	}
-});
-
 describe("when there are some blogs saved or new blogs are made", () => {
+	beforeEach(async () => {
+		await Blog.deleteMany({});
+		const users = await usersInDb();
+
+		for (let blog of initialBlogs) {
+			let blogObject = new Blog({...blog, user: users[0].id});
+			await blogObject.save();
+		}
+	});
+	const loginAndGetToken = async () => {
+		const response = await api.post("/api/login").send({
+			username: "root",
+			password: "sekret",
+		});
+		return `Bearer ${response.body.token}`;
+	};
 	test("blogs are returned as json", async () => {
 		await api
 			.get(baseUrl)
@@ -40,6 +46,7 @@ describe("when there are some blogs saved or new blogs are made", () => {
 	});
 
 	test("a valid blog has been created", async () => {
+		const token = await loginAndGetToken();
 		const newBlog = {
 			title: "Understanding Full stack development",
 			author: "Parth Kalra",
@@ -49,34 +56,28 @@ describe("when there are some blogs saved or new blogs are made", () => {
 
 		await api
 			.post(baseUrl)
+			.set("Authorization", token)
 			.send(newBlog)
 			.expect(201)
 			.expect("Content-Type", /application\/json/);
 
 		const response = await blogsInDb();
 
-		const contents = response.map((r) => {
-			return {
-				title: r.title,
-				author: r.author,
-				url: r.url,
-				likes: r.likes,
-			};
-		});
-
 		assert.strictEqual(response.length, initialBlogs.length + 1);
 		assert(
-			contents.some(
-				(r) =>
-					r.title === newBlog.title &&
-					r.author === newBlog.author &&
-					r.url === newBlog.url &&
-					r.likes === newBlog.likes
+			response.find(
+				(blog) =>
+					blog.title === newBlog.title &&
+					blog.author === newBlog.author &&
+					blog.url === newBlog.url &&
+					blog.likes === newBlog.likes
 			)
 		);
 	});
 
 	test("likes defaults to zero if missing from request", async () => {
+		const token = await loginAndGetToken();
+
 		const newBlog = {
 			title: "Demo Title",
 			author: "Demo User",
@@ -85,6 +86,7 @@ describe("when there are some blogs saved or new blogs are made", () => {
 
 		await api
 			.post(baseUrl)
+			.set("Authorization", token)
 			.send(newBlog)
 			.expect(201)
 			.expect("Content-Type", /application\/json/);
@@ -94,6 +96,7 @@ describe("when there are some blogs saved or new blogs are made", () => {
 	});
 
 	test("Server throws error of 400 on absence of title or url", async () => {
+		const token = await loginAndGetToken();
 		const blogWithoutUrl = {
 			title: "Demo Title 1",
 			author: "Demo User",
@@ -103,11 +106,20 @@ describe("when there are some blogs saved or new blogs are made", () => {
 			url: "https://example.com/demo-blog",
 		};
 
-		await api.post(baseUrl).send(blogWithoutUrl).expect(400);
-		await api.post(baseUrl).send(blogWithoutTitle).expect(400);
+		await api
+			.post(baseUrl)
+			.set("Authorization", token)
+			.send(blogWithoutUrl)
+			.expect(400);
+		await api
+			.post(baseUrl)
+			.set("Authorization", token)
+			.send(blogWithoutTitle)
+			.expect(400);
 	});
 
 	test("deletion works as intended", async () => {
+		const token = await loginAndGetToken();
 		const blogToBeDeleted = {
 			title: "Understanding JavaScript Closures",
 			author: "Jane Doe",
@@ -117,7 +129,7 @@ describe("when there are some blogs saved or new blogs are made", () => {
 		const blogListAtStart = await blogsInDb();
 		const blog = await Blog.findOne(blogToBeDeleted);
 
-		await api.delete(`${baseUrl}/${blog.id}`);
+		await api.delete(`${baseUrl}/${blog.id}`).set("Authorization", token);
 
 		const blogListAtEnd = await blogsInDb();
 		const searchForDeletedBlog = await Blog.findOne(blog);
@@ -127,6 +139,7 @@ describe("when there are some blogs saved or new blogs are made", () => {
 	});
 
 	test("update works as intended", async () => {
+		const token = await loginAndGetToken();
 		const blogToBeUpdated = {
 			title: "A Guide to Modern CSS Layouts",
 			author: "John Smith",
@@ -144,19 +157,24 @@ describe("when there are some blogs saved or new blogs are made", () => {
 
 		await api
 			.put(`${baseUrl}/${blog.id}`)
+			.set("Authorization", token)
 			.send(blogToBeUpdated)
 			.expect(200)
 			.expect("Content-Type", /application\/json/);
 
-		const searchForUpdatedBlog = await api.get(`${baseUrl}/${blog.id}`);
+		const searchForUpdatedBlog = await Blog.findById(blog.id);
 
-		assert.deepStrictEqual(searchForUpdatedBlog.body, {
-			...blogToBeUpdated,
-			id: blog.id,
-		});
+		assert.deepStrictEqual(
+			{
+				title: searchForUpdatedBlog.title,
+				author: blogToBeUpdated.author,
+				url: blogToBeUpdated.url,
+				likes: searchForUpdatedBlog.likes,
+			},
+			blogToBeUpdated
+		);
 	});
-});
-
-after(async () => {
-	await mongoose.connection.close();
+	after(async () => {
+		await mongoose.connection.close();
+	});
 });
